chore(docus): add intro flow nodes
