+++ docs2/components/Flows/Home/Intro.vue
@@ -1,8 +1,8 @@
 <script lang="ts" setup>
-import { Handle, Position, VueFlow, useVueFlow } from '@vue-flow/core'
+import { VueFlow, useVueFlow } from '@vue-flow/core'
 import { breakpointsTailwind, useBreakpoints } from '@vueuse/core'
 import { Background } from '@vue-flow/background'
-import Heart from '~icons/mdi/heart'
+import { CreditNode, DocsNode, ExamplesNode, IntroNode } from './Nodes'
 
 const breakpoints = useBreakpoints(breakpointsTailwind)
 
@@ -21,6 +21,13 @@ onMounted(() => {
     attributeFilter: ['class'],
   })
 })
+
+const nodeTypes = {
+  intro: markRaw(IntroNode),
+  examples: markRaw(ExamplesNode),
+  documentation: markRaw(DocsNode),
+  acknowledgement: markRaw(CreditNode),
+}
 
 const initialEdges = [
   {
@@ -49,11 +56,12 @@ const initialEdges = [
   },
 ]
 const { getNodes, findNode, setEdges, updateNodeInternals, dimensions, onNodesInitialized } = useVueFlow({
+  nodeTypes,
   nodes: [
-    { id: 'intro', type: 'box', position: { x: 0, y: 0 } },
-    { id: 'examples', type: 'box', position: { x: -50, y: 400 } },
-    { id: 'documentation', type: 'box', position: { x: 300, y: 400 } },
-    { id: 'acknowledgement', type: 'box', position: { x: 150, y: 500 } },
+    { id: 'intro', type: 'intro', position: { x: 0, y: 0 } },
+    { id: 'examples', type: 'examples', position: { x: -50, y: 400 } },
+    { id: 'documentation', type: 'documentation', position: { x: 300, y: 400 } },
+    { id: 'acknowledgement', type: 'acknowledgement', position: { x: 150, y: 500 } },
   ],
   edges: initialEdges,
   elementsSelectable: true,
@@ -175,150 +183,15 @@ function setElements() {
 }
 const { stop } = useResizeObserver(el, useDebounceFn(setElements, 5))
 onBeforeUnmount(stop)
-
-const scrollTo = () => {
-  const el = document.getElementById('acknowledgement')
-  if (el) {
-    el.scrollIntoView({ behavior: 'smooth' })
-  }
-}
 </script>
 
 <template>
   <VueFlow ref="el">
     <Background variant="lines" :pattern-color="dark ? '#ffffff' : '#000000'" :size="0.7" :gap="100" />
-
-    <template #node-box="props">
-      <template v-if="props.id === 'intro'">
-        <div class="box max-w-[500px]">
-          <div class="intro px-4 py-2 shadow-lg rounded-md border-2 border-solid border-black">
-            <div class="font-mono flex flex-col gap-4 p-4 items-center text-center">
-              <h1 class="text-2xl lg:text-4xl !my-0 !pt-0 font-bold">Vue Flow</h1>
-              <h2 class="text-lg lg:text-xl font-normal !border-0">
-                The customizable Vue 3 component bringing interactivity to flowcharts and graphs.
-              </h2>
-            </div>
-            <Handle id="a" type="source" :position="Position.Bottom" />
-          </div>
-        </div>
-      </template>
-      <template v-else-if="props.id === 'documentation'">
-        <div class="flex">
-          <a class="intro-link group bg-[#f15a16]" href="/guide/"> Read The Documentation </a>
-        </div>
-        <Handle type="target" :position="Position.Top" />
-        <Handle class="block md:hidden" type="source" :position="Position.Bottom" />
-      </template>
-      <template v-else-if="props.id === 'examples'">
-        <div class="flex">
-          <a class="intro-link group bg-[#ef467e]" href="/examples/"> Check The Examples </a>
-        </div>
-        <Handle type="target" :position="Position.Top" />
-        <Handle class="block md:hidden" type="source" :position="Position.Bottom" />
-      </template>
-      <template v-else-if="props.id === 'acknowledgement'">
-        <div class="flex" @click="scrollTo">
-          <button class="intro-link group bg-sky-500"><Heart class="text-red-500" />Acknowledgement</button>
-        </div>
-        <Handle type="target" :position="Position.Top" />
-      </template>
-    </template>
   </VueFlow>
 </template>
 
 <style>
-.checker-gb {
-  animation: fill-green-blue var(--animation-duration) alternate infinite;
-}
-
-.checker-gb path {
-  animation: fill-green-blue var(--animation-duration) alternate infinite;
-}
-
-.checker-op {
-  animation: fill-orange-purple var(--animation-duration) alternate infinite;
-}
-
-.checker-op path {
-  animation: fill-orange-purple var(--animation-duration) alternate infinite;
-}
-
-.checker-yg {
-  animation: fill-yellow-green var(--animation-duration) alternate infinite;
-}
-
-.checker-yg path {
-  animation: fill-yellow-green var(--animation-duration) alternate infinite;
-}
-
-.checker-ss {
-  animation: fill-sky-red var(--animation-duration) alternate infinite;
-}
-
-.checker-ss path {
-  animation: fill-sky-red var(--animation-duration) alternate infinite;
-}
-
-@keyframes fill-green-blue {
-  0% {
-    @apply fill-transparent;
-  }
-  35% {
-    @apply fill-green-500/50;
-  }
-  65% {
-    @apply fill-transparent;
-  }
-  100% {
-    @apply fill-blue-500/50;
-  }
-}
-
-@keyframes fill-orange-purple {
-  0% {
-    @apply fill-transparent;
-  }
-  35% {
-    @apply fill-orange-500/50;
-  }
-  65% {
-    @apply fill-transparent;
-  }
-  100% {
-    @apply fill-purple-500/50;
-  }
-}
-
-@keyframes fill-yellow-green {
-  0% {
-    @apply fill-transparent;
-  }
-  35% {
-    @apply fill-yellow-500/50;
-  }
-  65% {
-    @apply fill-transparent;
-  }
-  100% {
-    @apply fill-green-500/50;
-  }
-}
-
-@keyframes fill-sky-red {
-  0% {
-    @apply fill-transparent;
-  }
-  35% {
-    @apply fill-sky-500/50;
-  }
-  65% {
-    @apply fill-transparent;
-  }
-  100% {
-    @apply fill-red-500/50;
-  }
-}
-
 .intro {
   @apply cursor-pointer
   bg-green-500
