+++ docs2/components/Flows/Home/Nodes/index.ts
@@ -1,0 +1,4 @@
+export { default as CreditNode } from './CreditNode.vue'
+export { default as DocsNode } from './DocsNode.vue'
+export { default as ExamplesNode } from './ExamplesNode.vue'
+export { default as IntroNode } from './IntroNode.vue'
